Refactor: token service로 token 처리 내용 이동
